requirements.txt
snowflake-snowpark-python==1.48.1
python-dotenv==1.2.2
pyyaml==6.0.2
pytest==9.0.2
